+++ tools/cookiecutter-framework-project/{{cookiecutter.project_name}}/extensions/plugins/{{ cookiecutter.integration_name }}_save_to_temp_finalizer.py
@@ -14,7 +14,7 @@ class {{ cookiecutter.integration_name.capitalize() }}SaveToTempPlugin(BasePlugi
         Makes sure that the current opened scene is saved to a temp file so
         prevents it to be overriden.
         '''
-        # scene_type = '.mb'
+        scene_type = '.mb'
 
         try:
             # Save file to a temp file