+++ tools/cookiecutter-framework-project/{{cookiecutter.project_name}}/extensions/plugins/{{ cookiecutter.integration_name }}_scene_opener.py
@@ -25,6 +25,7 @@ class {{ cookiecutter.integration_name.capitalize() }}SceneOpenerPlugin(BasePlug
 
         try:
             # TODO: Open file in DCC
+            pass
         except Exception as error:
             raise PluginExecutionError(
                 f"Couldn't open the given path. Error: {error}"