+++ libs/qt-style/release_notes.md
@@ -1,6 +1,12 @@
 # ftrack Framework QT Style library release Notes
 
-## upcoming
+## v2.2.1
+2024-05-02
+
+* [fix] Replace resource.py import to be PySide2 and PySide6 compatible.
+
+## v2.2.0
+2024-05-02
 
 * [new] PySide6 support, through build script.
 * [new] PySide2 support, through build script.