+++ projects/framework-nuke/release_notes.md
@@ -1,8 +1,16 @@
 # ftrack Framework Nuke integration release Notes
 
-## Upcoming
+## v24.5.0
+2024-05-03
 
 * [changed] Replace Qt.py imports to PySide2 and PySide6 on widgets.
+
+
+## v24.4.1
+2024-04-30
+
+* [fix] Launcher; Properly escaped version expressions.
+
 
 ## v24.4.0
 2024-04-02