+++ projects/framework-photoshop/release_notes.md
@@ -1,7 +1,11 @@
 # ftrack Framework Photoshop integration release Notes
 
-## Upcoming
+## 24.5.0
+2024-05-13
 
+* [fix] Fixed bug where document could not be opened on Windows.
+* [changed] Fixed styling bug in panel causing footer to have different grey background.
+* [changed] Aligned PS with changes from Premiere, fix CI.
 * [changed] Moved extendscript jsx resources to extensions js folder, added PS specific bootstrap extension.
 * [changed] Refactored the code to enable code sharing with other Adobe CEP integrations.
 * [changed] Replace Qt.py imports to PySide2 and PySide6 on widgets.